+++ cli/admin/src/test/java/io/pravega/cli/admin/utils/TestUtils.java
@@ -317,173 +317,6 @@ public final class TestUtils {
     }
 
     /**
-     * Sets up a new BookKeeper & ZooKeeper.
-     */
-    public static class BookKeeperRunner implements AutoCloseable {
-        @Getter
-        private final AtomicReference<CuratorFramework> zkClient = new AtomicReference<>();
-        @Getter
-        private final int bkPort;
-        private final BookKeeperServiceRunner bookKeeperServiceRunner;
-        @Getter
-        private final AtomicReference<BookKeeperConfig> bkConfig = new AtomicReference<>();
-        private final AtomicReference<BookKeeperServiceRunner> bkService = new AtomicReference<>();
-        @Getter
-        private final String ledgerPath;
-        @Getter
-        private final String logMetaNamespace;
-        @Getter
-        private final String baseNamespace;
-
-        BookKeeperRunner(int instanceId, int bookieCount) throws Exception {
-            this.ledgerPath = "/pravega/bookkeeper/ledgers" + instanceId;
-            this.bkPort = io.pravega.test.common.TestUtils.getAvailableListenPort();
-            val bookiePorts = new ArrayList<Integer>();
-            for (int i = 0; i < bookieCount; i++) {
-                bookiePorts.add(io.pravega.test.common.TestUtils.getAvailableListenPort());
-            }
-            this.bookKeeperServiceRunner = BookKeeperServiceRunner.builder()
-                    .startZk(true)
-                    .zkPort(bkPort)
-                    .ledgersPath(ledgerPath)
-                    .bookiePorts(bookiePorts)
-                    .build();
-            try {
-                this.bookKeeperServiceRunner.startAll();
-            } catch (Exception e) {
-                log.error("Exception occurred while starting bookKeeper service.", e);
-                this.close();
-                throw e;
-            }
-            this.bkService.set(this.bookKeeperServiceRunner);
-
-            // Create a ZKClient with a unique namespace.
-            this.baseNamespace = "pravega" + instanceId;
-            this.zkClient.set(CuratorFrameworkFactory
-                    .builder()
-                    .connectString("localhost:" + bkPort)
-                    .namespace(baseNamespace)
-                    .retryPolicy(new ExponentialBackoffRetry(1000, 10))
-                    .build());
-
-            this.zkClient.get().start();
-
-            logMetaNamespace = "segmentstore/containers" + instanceId;
-            this.bkConfig.set(BookKeeperConfig
-                    .builder()
-                    .with(BookKeeperConfig.ZK_ADDRESS, "localhost:" + bkPort)
-                    .with(BookKeeperConfig.ZK_METADATA_PATH, logMetaNamespace)
-                    .with(BookKeeperConfig.BK_LEDGER_PATH, ledgerPath)
-                    .build());
-        }
-
-        @Override
-        public void close() throws Exception {
-            val process = this.bkService.getAndSet(null);
-            if (process != null) {
-                process.close();
-            }
-
-            val bk = this.bookKeeperServiceRunner;
-            if (bk != null) {
-                bk.close();
-            }
-
-            val zkClient = this.zkClient.getAndSet(null);
-            if (zkClient != null) {
-                zkClient.close();
-            }
-        }
-    }
-
-    /**
-     * Creates a segment store.
-     */
-    public static class SegmentStoreRunner implements AutoCloseable {
-        private final int servicePort = io.pravega.test.common.TestUtils.getAvailableListenPort();
-        @Getter
-        private final int adminPort = io.pravega.test.common.TestUtils.getAvailableListenPort();
-        @Getter
-        private final ServiceBuilder serviceBuilder;
-        private final PravegaConnectionListener server;
-        private final StreamSegmentStore streamSegmentStore;
-        private final TableStore tableStore;
-
-        private AdminConnectionListener adminServer = null;
-
-        SegmentStoreRunner(StorageFactory storageFactory, BookKeeperLogFactory dataLogFactory, int containerCount, boolean enableAdminGateway)
-                throws DurableDataLogException {
-            ServiceBuilderConfig.Builder configBuilder = ServiceBuilderConfig
-                    .builder()
-                    .include(ServiceConfig.builder()
-                            .with(ServiceConfig.CONTAINER_COUNT, containerCount)
-                            .with(ServiceConfig.STORAGE_LAYOUT, StorageLayoutType.CHUNKED_STORAGE)
-                            .with(ServiceConfig.STORAGE_IMPLEMENTATION, ServiceConfig.StorageType.FILESYSTEM.name()))
-                    .include(WriterConfig.builder()
-                            .with(WriterConfig.MIN_READ_TIMEOUT_MILLIS, 100L)
-                            .with(WriterConfig.MAX_READ_TIMEOUT_MILLIS, 500L)
-                    );
-            if (storageFactory != null) {
-                if (dataLogFactory != null) {
-                    this.serviceBuilder = ServiceBuilder.newInMemoryBuilder(configBuilder.build())
-                            .withStorageFactory(setup -> storageFactory)
-                            .withDataLogFactory(setup -> dataLogFactory);
-                } else {
-                    this.serviceBuilder = ServiceBuilder.newInMemoryBuilder(configBuilder.build())
-                            .withStorageFactory(setup -> storageFactory);
-                }
-            } else {
-                this.serviceBuilder = ServiceBuilder.newInMemoryBuilder(ServiceBuilderConfig.getDefaultConfig());
-            }
-            this.serviceBuilder.initialize();
-            this.streamSegmentStore = this.serviceBuilder.createStreamSegmentService();
-            this.tableStore = this.serviceBuilder.createTableStoreService();
-            this.server = new PravegaConnectionListener(false, servicePort, this.streamSegmentStore, this.tableStore,
-                    this.serviceBuilder.getLowPriorityExecutor());
-            this.server.startListening();
-            if (enableAdminGateway) {
-                this.adminServer = new AdminConnectionListener(false, false, "localhost", adminPort, this.streamSegmentStore,
-                        this.tableStore, new PassingTokenVerifier(), null, null, TLS_PROTOCOL_VERSION.getDefaultValue().split(","));
-                this.adminServer.startListening();
-            }
-        }
-
-        @Override
-        public void close() {
-            this.server.close();
-            if (adminServer != null) {
-                this.adminServer.close();
-            }
-            this.serviceBuilder.close();
-        }
-    }
-
-    /**
-     * Creates a controller instance and runs it.
-     */
-    public static class ControllerRunner implements AutoCloseable {
-        private final int controllerPort = io.pravega.test.common.TestUtils.getAvailableListenPort();
-        private final String serviceHost = "localhost";
-        private final ControllerWrapper controllerWrapper;
-        @Getter
-        private final Controller controller;
-        private final URI controllerURI = URI.create("tcp://" + serviceHost + ":" + controllerPort);
-
-        ControllerRunner(int bkPort, int servicePort, int containerCount) {
-            this.controllerWrapper = new ControllerWrapper("localhost:" + bkPort, false,
-                    controllerPort, serviceHost, servicePort, containerCount);
-            this.controllerWrapper.awaitRunning();
-            this.controller = controllerWrapper.getController();
-        }
-
-        @Override
-        public void close() throws Exception {
-            this.controller.close();
-            this.controllerWrapper.close();
-        }
-    }
-
-    /**
      * Creates a client to read and write events.
      */
     public static class ClientRunner implements AutoCloseable {
@@ -493,11 +326,11 @@ public final class TestUtils {
         @Getter
         private final ReaderGroupManager readerGroupManager;
 
-        public ClientRunner(ControllerRunner controllerRunner, String scope) {
+        public ClientRunner(InProcessServiceStarter.ControllerRunner controllerRunner, String scope) {
             this.connectionFactory = new SocketConnectionFactoryImpl(ClientConfig.builder()
-                    .controllerURI(controllerRunner.controllerURI).build());
-            this.clientFactory = new ClientFactoryImpl(scope, controllerRunner.controller, connectionFactory);
-            this.readerGroupManager = new ReaderGroupManagerImpl(scope, controllerRunner.controller, clientFactory);
+                    .controllerURI(controllerRunner.getControllerURI()).build());
+            this.clientFactory = new ClientFactoryImpl(scope, controllerRunner.getController(), connectionFactory);
+            this.readerGroupManager = new ReaderGroupManagerImpl(scope, controllerRunner.getController(), clientFactory);
         }
 
         @Override
@@ -505,60 +338,6 @@ public final class TestUtils {
             this.readerGroupManager.close();
             this.clientFactory.close();
             this.connectionFactory.close();
-        }
-    }
-
-    /**
-     * Creates a Pravega instance.
-     */
-    public static class PravegaRunner implements AutoCloseable {
-        private final int containerCount;
-        private final int bookieCount;
-        @Getter
-        private BookKeeperRunner bookKeeperRunner;
-        @Getter
-        private SegmentStoreRunner segmentStoreRunner;
-        @Getter
-        private ControllerRunner controllerRunner;
-
-        public PravegaRunner(int bookieCount, int containerCount) {
-            this.containerCount = containerCount;
-            this.bookieCount = bookieCount;
-        }
-
-        public void startBookKeeperRunner(int instanceId) throws Exception {
-            this.bookKeeperRunner = new BookKeeperRunner(instanceId, this.bookieCount);
-        }
-
-        public void startControllerAndSegmentStore(StorageFactory storageFactory, BookKeeperLogFactory dataLogFactory)
-                throws DurableDataLogException {
-            startControllerAndSegmentStore(storageFactory, dataLogFactory, false);
-        }
-
-        public void startControllerAndSegmentStore(StorageFactory storageFactory, BookKeeperLogFactory dataLogFactory, boolean enableAdminGateway)
-                throws DurableDataLogException {
-            this.segmentStoreRunner = new SegmentStoreRunner(storageFactory, dataLogFactory, this.containerCount, enableAdminGateway);
-            log.info("bk port to be connected = {}", this.bookKeeperRunner.bkPort);
-            this.controllerRunner = new ControllerRunner(this.bookKeeperRunner.bkPort, this.segmentStoreRunner.servicePort, containerCount);
-        }
-
-        public void shutDownControllerRunner() throws Exception {
-            this.controllerRunner.close();
-        }
-
-        public void shutDownSegmentStoreRunner() {
-            this.segmentStoreRunner.close();
-        }
-
-        public void shutDownBookKeeperRunner() throws Exception {
-            this.bookKeeperRunner.close();
-        }
-
-        @Override
-        public void close() throws Exception {
-            shutDownControllerRunner();
-            shutDownSegmentStoreRunner();
-            shutDownBookKeeperRunner();
         }
     }
 }
